+++ test/test_helper.rb
@@ -5,6 +5,7 @@ require_relative "../test/dummy/config/environment"
 ActiveRecord::Migrator.migrations_paths = [File.expand_path("../test/dummy/db/migrate", __dir__)]
 ActiveRecord::Migrator.migrations_paths << File.expand_path("../db/migrate", __dir__)
 require "rails/test_help"
+require "minitest/mock"
 
 # Load fixtures from the engine
 if ActiveSupport::TestCase.respond_to?(:fixture_paths=)
@@ -21,5 +22,14 @@ class ActiveSupport::TestCase
     yield
   ensure
     Rails.cache = previous_cache
+  end
+
+  def with_stubbed_job_fetch(job_class, body)
+    job = job_class.new
+    job.stub(:fetch, body) do
+      job_class.stub(:new, job) do
+        yield
+      end
+    end
   end
 end